requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
yt-dlp==2023.11.16
openai-whisper==20231117
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy<2
jinja2
requests